+++ efinance/stock/__init__.py
@@ -10,7 +10,8 @@ from .getter import (
     get_all_company_performance,
     get_latest_holder_number,
     get_daily_billboard,
-    get_members)
+    get_members,
+    get_latest_ipo_info)
 __all__ = ['get_history_bill',
            'get_today_bill',
            'get_latest_quote',
@@ -22,4 +23,5 @@ __all__ = ['get_history_bill',
            'get_all_company_performance',
            'get_latest_holder_number',
            'get_daily_billboard',
-           'get_members']
+           'get_members',
+           'get_latest_ipo_info']